Cleanup: put hotstuffProposeHandler to a separate package.
