+++ dependencies/protocol.go
@@ -4,6 +4,7 @@ import (
 	"github.com/relab/hotstuff/modules"
 	"github.com/relab/hotstuff/network"
 	"github.com/relab/hotstuff/protocol/consensus"
+	"github.com/relab/hotstuff/protocol/hsproposehandler"
 	"github.com/relab/hotstuff/protocol/kauri"
 	"github.com/relab/hotstuff/protocol/proposer"
 	"github.com/relab/hotstuff/protocol/synchronizer"
@@ -58,7 +59,7 @@ func NewProtocol(
 		state,
 	)
 	var handler modules.ExtProposeHandler
-	// TODO(AlanRostem): avoid using runtime cfg to set kauri.
+	// TODO(AlanRostem): add module logic for this.
 	if depsCore.RuntimeCfg().KauriEnabled() {
 		handler = kauri.New(
 			depsCore.Logger(),
@@ -69,7 +70,7 @@ func NewProtocol(
 			sender.(*network.GorumsSender), // TODO(AlanRostem): avoid cast
 		)
 	} else {
-		handler = consensus.NewHotStuffProposeHandler(
+		handler = hsproposehandler.New(
 			depsCore.Logger(),
 			depsCore.RuntimeCfg(),
 			voter,
